+++ integration_tests/suite/base/test_device_associations.py
@@ -1,18 +1,19 @@
 # -*- coding: utf-8 -*-
-# Copyright 2016-2017 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2016-2018 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0+
 
 from __future__ import unicode_literals
 
 from hamcrest import assert_that, has_entries
 
-
-from ..helpers import config
-from ..helpers import errors as e
-from ..helpers import fixtures
-from ..helpers import associations as a
-from ..helpers import helpers as h
 from . import mocks, provd, db, confd
+from ..helpers import (
+    associations as a,
+    config,
+    errors as e,
+    fixtures,
+    helpers as h,
+)
 
 
 @fixtures.user()